requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
scikit-learn
numpy
matplotlib
konlpy
keybert
torch
transformers
sentencepiece
